+++ src/lib/helpers/string.helper.ts
@@ -1,7 +1,5 @@
-import { isNullOrUndefined } from 'util';
-
 export class StringHelper {
     public static isNullUndefinedOrEmpty(str: string): boolean {
-        return isNullOrUndefined(str) || str.length === 0;
+        return str === null || str === undefined || str.length === 0;
     }
 }