+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Features
 
 -   **dependencies** updated all `jasmine` and `karma` packages to the latest compatible versions.
+-   **string-helper** changed deprecated util function to plain js in method `isNullUndefinedOrEmpty`.
 
 # 6.0.0-alpha.5 (21.12.2020)